Move geometry decoding logic to its own file
